fix reverse motors

+++ arduino/headers.h
@@ -4,7 +4,7 @@ void setupLeds();
 void ledController();
 void setLeds(uint8_t pwm, uint32_t trans_t);
 void setupMotors();
-void setMotors(uint8_t motor, int8_t pwm, uint32_t trans_t);
+void setMotors(uint8_t motor, int pwm, uint32_t trans_t);
 void motorController();
 void motor2Controller();
 void stopLeds();
